Removing AutoMapper

+++ FN.Testing.Business/Services/UploadDataService.cs
@@ -1,57 +1,47 @@
-﻿using AutoMapper;
-using FN.Testing.Business.Contract.Abstractions;
+﻿using FN.Testing.Business.Contract.Abstractions;
 using FN.Testing.Business.Contract.Entities;
 using FN.Testing.DataLayer.Contract.Abstractions;
-using FN.Testing.DataLayer.Contract.Tables;
 using FN.Testing.Functions;
-using Microsoft.AspNetCore.StaticFiles;
 using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
+using FN.Testing.Common.Mappings;
 
 namespace FN.Testing.Business.Services
 {
     public class UploadDataService : IUploadDataService
     {
         private readonly IUploadRepository _repository;
-        private readonly IMapper _mapper;
-        public int Instance;
-        public static int InstanceCount;
 
         public UploadDataService(
-            IUploadRepository repository,
-            IMapper mapper)
+            IUploadRepository repository)
         {
-            InstanceCount++;
-            Instance = InstanceCount;
             _repository = repository ?? throw new System.ArgumentNullException(nameof(repository));
-            _mapper = mapper ?? throw new System.ArgumentNullException(nameof(mapper));
         }
         public async Task<UploadedEntity> GetUpload(int id, CancellationToken cancellationToken)
         {
             var row = await _repository.GetUploadById(id, cancellationToken);
-            return _mapper.Map<UploadedEntity>(row);
+            return row.ToUploadedEntity();
         }
         public async Task<IEnumerable<UploadedEntity>> GetUploads(CancellationToken cancellationToken)
         {
             var rows = await _repository.GetUploads();
-            return _mapper.Map<IEnumerable<UploadedEntity>>(rows);
+            return rows.ToUploadedEntities();
         }
         public async Task DeleteUpload(int id, CancellationToken cancellationToken)
         {
             var file = await GetUpload(id, cancellationToken);
-            if (file != null)
+            if (file != null && new Uploader().DeleteFile(string.Concat(file.FileName, file.Extension)))
             {
-                if (new Uploader().DeleteFile(string.Concat(file.FileName, file.Extension)))
-                    await _repository.DeleteUpload(id, cancellationToken);
+                await _repository.DeleteUpload(id, cancellationToken);
             }            
         }
         public async Task<UploadedEntity> AddUpload(UploadEntity entity, CancellationToken cancellationToken)
         {
             var uploadPath = new Uploader().UploadFile(entity.File);
-            var inputEt = _mapper.Map<Upload>(entity);
+            var inputEt = entity.ToUpload();
             inputEt.UploadDate = DateTime.Now;
             inputEt.Id = 0;
             inputEt.FileName = Path.GetFileNameWithoutExtension(uploadPath);
